pyproject.toml
[project]
name = "autonomous-incident-response-agent"
version = "0.1.0"
description = "Autonomous DevOps incident response — RAG, agent, API (phased build)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.0",
    "openai>=1.40.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "langgraph>=1.1.4",
    "langchain-openai>=1.1.12",
    "langchain-core>=1.2.24",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.6.0", "httpx>=0.27.0"]
ui = ["gradio>=5.0.0"]

[project.scripts]
rag-build = "app.rag.cli:main_build"
rag-query = "app.rag.cli:main_query"
triage = "app.agent.cli:main"
triage-eval = "app.eval.cli:main"
serve-api = "app.api.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore:builtin type SwigPyPacked:DeprecationWarning",
    "ignore:builtin type SwigPyObject:DeprecationWarning",
    "ignore:builtin type swigvarlink:DeprecationWarning",
]
